Refactor sign up page with tailwindcss.
